+++ jdaviz/configs/specviz2d/plugins/spectral_extraction/spectral_extraction.py
@@ -74,6 +74,9 @@ class SpectralExtraction(PluginTemplateMixin):
     ext_dataset_items = List().tag(sync=True)
     ext_dataset_selected = Unicode().tag(sync=True)
 
+    ext_type_items = List(['Boxcar', 'Optimal']).tag(sync=True)
+    ext_type_selected = Unicode('Boxcar').tag(sync=True)
+
     ext_width = IntHandleEmpty(0).tag(sync=True)
 
     ext_specreduce_err = Unicode().tag(sync=True)
@@ -363,7 +366,7 @@ class SpectralExtraction(PluginTemplateMixin):
         self.active_step = 'bg'
         self._update_plugin_marks()
 
-    @observe('ext_dataset_selected', 'ext_width', 'active_step')
+    @observe('ext_dataset_selected', 'ext_type_selected', 'ext_width', 'active_step')
     def _interaction_in_ext_step(self, event={}):
         if not self.plugin_opened or not self._do_marks:
             return
@@ -381,10 +384,14 @@ class SpectralExtraction(PluginTemplateMixin):
             self.marks['trace'].update_xy(xs,
                                           trace.trace)
             self.marks['trace'].line_style = 'dashed'
-            self.marks['ext_lower'].update_xy(xs,
-                                              trace.trace-self.ext_width/2)
-            self.marks['ext_upper'].update_xy(xs,
-                                              trace.trace+self.ext_width/2)
+            if self.ext_type_selected == 'Boxcar':
+                self.marks['ext_lower'].update_xy(xs,
+                                                  trace.trace-self.ext_width/2)
+                self.marks['ext_upper'].update_xy(xs,
+                                                  trace.trace+self.ext_width/2)
+            else:
+                for mark in ['ext_lower', 'ext_upper']:
+                    self.marks[mark].clear()
 
         self.active_step = 'ext'
         self._update_plugin_marks()
@@ -594,12 +601,13 @@ class SpectralExtraction(PluginTemplateMixin):
         ext : specreduce.extract.BoxcarExtract
             Extract object to import
         """
-        if not isinstance(ext, extract.BoxcarExtract):  # pragma: no cover
-            # TODO: add support for Optimal/Horne
-            raise TypeError("ext must be a specreduce.extract.BoxcarExtract object")
-
-        # TODO: should we detect/set the referenced dataset?
-        self.ext_width = ext.width
+        if isinstance(ext, extract.BoxcarExtract):
+            self.ext_type_selected = 'Boxcar'
+            self.ext_width = ext.width
+        elif isinstance(ext, extract.HorneExtract):
+            self.ext_type_selected = 'Horne'
+        else:  # pragma: no cover
+            raise TypeError("ext must be a specreduce.extract.BoxcarExtract or specreduce.extract.HorneExtract object")  # noqa
 
     def export_extract(self, **kwargs):
         """
@@ -610,9 +618,20 @@ class SpectralExtraction(PluginTemplateMixin):
             self.update_marks(step='ext')
 
         trace = self._get_ext_trace()
-        inp_sp2d = self._get_ext_input_spectrum().flux.value
+        inp_sp2d = self._get_ext_input_spectrum()
 
-        return extract.BoxcarExtract(inp_sp2d, trace, width=self.ext_width)
+        if self.ext_type_selected == 'Boxcar':
+            ext = extract.BoxcarExtract(inp_sp2d.data, trace, width=self.ext_width)
+        elif self.ext_type_selected == 'Optimal':
+            # TODO: convert uncertainty depending on type?
+            variance = inp_sp2d.uncertainty
+            if variance is None:
+                variance = np.ones_like(inp_sp2d.data)
+            ext = extract.OptimalExtract(inp_sp2d.data, trace, variance=variance)
+        else:
+            raise NotImplementedError(f"extraction type '{self.ext_type_selected}' not supported")  # noqa
+
+        return ext
 
     def export_extract_spectrum(self, add_data=False, **kwargs):
         """
@@ -626,6 +645,7 @@ class SpectralExtraction(PluginTemplateMixin):
         """
         extract = self.export_extract(**kwargs)
         spectrum = extract.spectrum
+
         # Specreduce returns a spectral axis in pixels, so we'll replace with input spectral_axis
         # NOTE: this is currently disabled until proper handling of axes-limit linking between
         # the 2D spectrum image (plotted in pixels) and a 1D spectrum (plotted in freq or